Added Travis Config & Grayson CLI Tests
Added `shelljs` dependency to run tests on `grayson-init.js` and
`grayson-gen.js` scripts. This tests the actual functionality of the
grayson package, rather than testing just the inner functionality of the
templating.

Added config file for Travis CI, to provide feedback for users/consumers
as to the current state of the package's test coverage.
⚡️ Changes to be committed:
-	new file:   .travis.yml
-	modified:   package-lock.json
-	modified:   package.json
-	new file:   tests/grayson.test.js
-	modified:   tests/nav-element.test.js